Move Protocol from resolver to proto
